[!!!][TASK] Improve CLi command and type hinting
This adds an artificial "lifetime" to the
event processing CLI command in order to not run
into typical long-running-PHP-processes problems
like timing out DB connections and so on.

You will need to daemonize the respective command
in order to respawn it whenever it quits intentionally.

This would be possible via ``daemon``, ``upstart`` or
``supervisord``.

This is considered breaking because you now need
to take care to respawn that process.

+++ Classes/Ag/Event/Command/EventCommandController.php
@@ -2,7 +2,10 @@
 namespace Ag\Event\Command;
 
 use Ag\Event\Domain\Model\DomainEvent;
+use Ag\Event\Domain\Model\StoredEvent;
+use Ag\Event\Exception\EventHandlingException;
 use Ag\Event\Service\EventService;
+use Pheanstalk\Job;
 use Pheanstalk\Pheanstalk;
 use TYPO3\Flow\Annotations as Flow;
 use TYPO3\Flow\Cli\CommandController;
@@ -12,6 +15,16 @@ use TYPO3\Flow\Log\SystemLoggerInterface;
  * @Flow\Scope("singleton")
  */
 class EventCommandController extends CommandController {
+
+	/**
+	 * The amount of job queries until the process quits and has to be respawned
+	 */
+	const LOOPS = 100;
+
+	/**
+	 * The job queue timeout
+	 */
+	const QUEUE_TIMEOUT = 20;
 
 	/**
 	 * @Flow\Inject
@@ -37,19 +50,46 @@ class EventCommandController extends CommandController {
 	 * @param string $key
 	 */
 	public function processCommand($key) {
-		while(TRUE) {
-			$this->systemLogger->log('Waiting for event in tube ' . $key, LOG_DEBUG);
+		$this->tellStatus('Event processor "%s" starting.', array($key));
+
+		$i = 0;
+		while($i < self::LOOPS) {
+			$i++;
 			$job = $this->pheanstalk
 				  ->watch($key)
 				  ->ignore('default')
-				  ->reserve();
+				  ->reserve(self::QUEUE_TIMEOUT);
 
-			/** @var $domainEvent DomainEvent */
-			$domainEvent = unserialize($job->getData());
-			$this->eventService->handleDomainEventByEventHandler($domainEvent, str_replace('_', '\\', $key));
+			if ($job instanceof Job) {
+				$this->tellStatus('Found job #%d in tube "%s".', array($job->getId(), $key));
 
-			$this->pheanstalk->delete($job);
+				/** @var $storedEvent StoredEvent */
+				$storedEvent = unserialize($job->getData());
+
+				$this->tellStatus('Attempt to handle event #%s.', array($storedEvent->getEventId()));
+
+				$this->eventService->handleDomainEventByEventHandler($storedEvent->getEvent(), str_replace('_', '\\', $key));
+
+				$this->pheanstalk->delete($job);
+				$this->tellStatus('Deleted job #%d in tube "%s".', array($job->getId(), $key));
+			}
 		}
+
+		if ($i >= self::LOOPS) {
+			$this->tellStatus('Event processor reached %d loops. Will restart.', array(self::LOOPS));
+		}
+
+		$this->tellStatus('Event processor exited intentionally.');
+	}
+
+	/**
+	 * @param string $message
+	 * @param array $arguments
+	 */
+	protected function tellStatus($message, array $arguments = NULL) {
+		$message = vsprintf($message, $arguments);
+		$this->systemLogger->log($message, LOG_INFO);
+		echo sprintf('%s: %s', date(\DateTime::ISO8601), $message) . PHP_EOL;
 	}
 
 }
